refactor(setup): isolate stable JSON presentation
